Project on Crime Data India NCRB 2022
This is a project to handle 'big data' from the public source of Indian Crime Data on the NCRB 2022 platform. Initially, it consisted of a large volume of raw data even in different formats. I primarily used Python for processing, visualize and statistical analysis. This project is focused on educational and technical learning.