A Project on Indian Crime Data (NCRB 2022)
This is a project to handle 'big data' from the public source of  Indian Crime Data on the NCRB 2022 platform. Initially, it consisted of a large volume of raw data even in different nature. I make it a well-structured data. I primarily used Python for processing, visualize and analysis. This project is focused on educational and technical learning.